Copy in metadata tests
